test(session): split interaction provider tests to satisfy file-size lint
`interaction_pipeline_tests.rs` hit 457 lines (cap 400) after the #929
model-resolution + AC#5 tests landed. Move both into a new
`interaction_pipeline_provider_tests.rs` sibling, reusing the shared
helpers — same pattern as the existing `_guard_tests.rs` split.
